componentised chatroom
